+++ goatgam/src/main/java/com/sparta/goatgam/domain/user/controller/UserController.java
@@ -1,14 +1,14 @@
-package com.sparta.goatgam.controller;
+package com.sparta.goatgam.domain.user.controller;
 
-import com.sparta.goatgam.dto.LoginRequestDto;
-import com.sparta.goatgam.dto.SignupRequestDto;
-import com.sparta.goatgam.dto.UserInfoDto;
-import com.sparta.goatgam.entity.User;
-import com.sparta.goatgam.entity.UserRoleEnum;
-import com.sparta.goatgam.jwt.JwtUtil;
-import com.sparta.goatgam.repository.UserRepository;
-import com.sparta.goatgam.security.UserDetailsImpl;
-import com.sparta.goatgam.service.UserService;
+import com.sparta.goatgam.domain.user.dto.LoginRequestDto;
+import com.sparta.goatgam.domain.user.dto.SignupRequestDto;
+import com.sparta.goatgam.domain.user.dto.UserInfoDto;
+import com.sparta.goatgam.domain.user.entity.User;
+import com.sparta.goatgam.domain.user.entity.UserRoleEnum;
+import com.sparta.goatgam.global.jwt.jwt.JwtUtil;
+import com.sparta.goatgam.domain.user.repository.UserRepository;
+import com.sparta.goatgam.global.security.security.UserDetailsImpl;
+import com.sparta.goatgam.domain.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,9 +17,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.GetMapping;
