+++ goatgam/src/main/java/com/sparta/goatgam/domain/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.goatgam.repository;
+package com.sparta.goatgam.domain.user.repository;
 
 
-import com.sparta.goatgam.entity.User;
+import com.sparta.goatgam.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
