+++ goatgam/src/main/java/com/sparta/goatgam/global/jwt/jwt/JwtUtil.java
@@ -1,6 +1,6 @@
-package com.sparta.goatgam.jwt;
+package com.sparta.goatgam.global.jwt.jwt;
 
-import com.sparta.goatgam.entity.UserRoleEnum;
+import com.sparta.goatgam.domain.user.entity.UserRoleEnum;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import io.jsonwebtoken.*;
